split evaluator and stepper
