+++ project-cis5520.cabal
@@ -42,6 +42,7 @@ library
   exposed-modules:
       Lib
       LuStepper 
+      LuEvaluator
       LuParser
       LuSyntax 
       State
